requirements.txt
fastapi==0.116.1
SQLAlchemy[asyncio]==2.0.43
asyncpg==0.30.0
python-dotenv==1.1.1
uvicorn==0.35.0
alembic==1.16.5
psycopg2-binary==2.9.10